python3-numpy: don't do out-of-tree builds
The source generation in numpy appears to be fragile and silently fails
when an out-of-tree build is done with setuptools.  The result is that
importing the numpy library results in failing linkage.

We should move numpy to build with mesonpy, but until that is done we
can simply reset B to be at the default location under the source tree.

(From OE-Core rev: ec2fd11153eca6f622cd5674b6f1ce21db99b1d1)

Signed-off-by: Ross Burton <ross.burton@arm.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>